refactor: moves material validation to a middleware
